chore: move `enum` `null` and `object` outside the `lib`.
