+++ src/enum/enum-state.abstract.ts
@@ -1,5 +1,5 @@
 // Abstract.
-import { State } from "../state.abstract";
+import { State } from "../lib/state.abstract";
 /**
  * @description Handles the `enum` type state.
  * @export
