+++ src/enum/enum.class.ts
