+++ src/object/boolean-action-object-state.abstract.ts
