+++ src/object/boolean-object-state.abstract.ts
