+++ src/object/index.ts
