+++ src/object/named-object-state.abstract.ts
