+++ src/object/object-state.abstract.ts
